pyproject.toml
# Copyright 2023 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#      https://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[project]
name = "xpk"
dynamic = ["version"]
authors = [
  { name="XPK team", email="xpk-code-reviewers@google.com" },
]
description = "xpk helps Cloud developers to orchestrate training jobs on accelerators on GKE."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
keywords = []

# pip dependencies installed with `pip install -e .`
dependencies = [
  "cloud-accelerator-diagnostics==0.1.1",
  "tabulate==0.9.0",
  "ruamel.yaml==0.18.10",
  "pyyaml==6.0.2",
  "docker==7.1.0",
  "kubernetes==31.0.0",
  "google-cloud==0.34.0",
  "google-api-core==2.24.1",
  "packaging==24.2",
  "google-cloud-filestore==1.12.0",
  "google-cloud-storage==2.19.0"
]

[project.urls]
"Homepage" = "https://github.com/google/xpk"
"Bug Tracker" = "https://github.com/google/xpk/issues"

[build-system]
requires = ["setuptools>=61.0", "setuptools-git-versioning"]
build-backend = "setuptools.build_meta"

[project.scripts]
xpk = "xpk.main:main"

[project.optional-dependencies]
# Development deps (unittest, linting, formating,...)
# Installed through `pip install .[dev]`
dev = [
    "pyink==24.3.0",
    "pylint>=2.6.0",
    "pre-commit",
    "pytest",
    "docker==7.1.0"
]

[tool.setuptools]
packages = ["xpk", "xpk.parser", "xpk.core", "xpk.commands", "xpk.api", "xpk.templates", "xpk.utils", "xpk.core.blueprint", "xpk.core.remote_state", "xpk.core.workload_decorators"]
package-dir = {"" = "src"}
package-data = {"xpk.api" = ["storage_crd.yaml"], "xpk.templates" = ["storage.yaml"]}

[tool.setuptools-git-versioning]
enabled = true
dev_template = "{tag}+{sha}"
dirty_template = "{tag}+{sha}.dirty"

[tool.pyink]
# Formatting configuration to follow Google style-guide.
line-length = 80
unstable = true
pyink-indentation = 2
pyink-use-majority-quotes = true
